requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
dj-database-url==2.0.0
django-taggit==4.0.0
django==4.2.2
pyyaml==6.0
sqlparse==0.4.4; python_version >= '3.5'
thttp==1.2.0
typing-extensions==4.7.0; python_version >= '3.7'
